qt6: migration step 7 remaining build fixes (100% compilation)
Fix all remaining Qt6 compilation blockers discovered after step 6,
including QString comparison ambiguities, missing includes, JS engine
syntax, nullptr guards, and disabling M2K package (deferred to refactor).

+++ packages/extproc/plugins/extprocplugin/include/extprocplugin/controller/extprocplotinfo.h
@@ -85,7 +85,7 @@ public:
 
 	bool isValid() const
 	{
-		if(id < 0 || title.isEmpty() || xLabel.isEmpty() || yLabel.isEmpty() || type == INVALID_TYPE) {
+		if(id.toInt() < 0 || title.isEmpty() || xLabel.isEmpty() || yLabel.isEmpty() || type == INVALID_TYPE) {
 			return false;
 		}
